feat: Add VPC module with main.tf, outputs.tf, and variables.tf
This commit adds the VPC module to the repository. The module includes the main.tf file, which defines the AWS VPC resource with a specified CIDR block and tags. It also includes the outputs.tf file, which outputs the ID of the VPC. Lastly, it includes the variables.tf file, which defines the vpc_cidr variable for specifying the CIDR block of the VPC.